IDE-2054 fix test case error
